moves profile meta detection back to profilebuilder
